shape_cam/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
pyyaml>=5.4.0
opencv-python # Or Pillow for image I/O
matplotlib # For visualization (optional)
# Add other specific dependencies as needed